+++ src/test/java/org/arkecosystem/crypto/transactions/builder/ValidatorRegistrationBuilderTest.java
@@ -17,10 +17,10 @@ public class ValidatorRegistrationBuilderTest extends AbstractTest {
 
         ValidatorRegistrationBuilder builder =
                 new ValidatorRegistrationBuilder()
-                        .gasPrice(((Number) data.get("gasPrice")).intValue())
+                        .gasPrice(((Number) data.get("gasPrice")).longValue())
                         .nonce(Long.parseLong(data.get("nonce").toString()))
                         .network(((Number) data.get("network")).intValue())
-                        .gasLimit(((Number) data.get("gasLimit")).intValue())
+                        .gasLimit(((Number) data.get("gasLimit")).longValue())
                         .validatorPublicKey(
                                 "a08058db53e2665c84a40f5152e76dd2b652125a6079130d4c315e728bcf4dd1dfb44ac26e82302331d61977d3141118")
                         .recipientAddress((String) data.get("recipientAddress"))
